add shared messages
